+++ src/api/v1/otp/otp.service.ts
@@ -1,14 +1,15 @@
 import {
   type GetOtp,
-  NotificationType,
   type OtpRecord,
   type SendOtpRequest,
   type SendOtpResponse,
-  UserActionType,
   type VerifyOtpRequest,
-  notificationToActionMap,
+  emailNotificationEvents,
+  otpEvents,
   otpSchema,
   sendOtpRequestSchema,
+  sendOtpToActionMap,
+  userActions,
   verifyOtpRequestSchema,
 } from "@ansospace/types";
 import { formatDuration, intervalToDuration, isPast } from "date-fns";
@@ -23,40 +24,35 @@ import { TokenService } from "../token/token.service.js";
 import { OtpDAL } from "./otp.dal.js";
 
 export class OtpService {
-  public static async sendOtp(otpEvents: SendOtpRequest): Promise<{ message: string } & SendOtpResponse> {
-    const validOtpEvent = sendOtpRequestSchema.parse(otpEvents);
-    const { otpType } = validOtpEvent;
+  public static async sendOtp(otpRequest: SendOtpRequest): Promise<{ message: string } & SendOtpResponse> {
+    const { eventType, email } = sendOtpRequestSchema.parse(otpRequest);
 
     const otp = generateOTP();
     const otpTTLDuration = intervalToDuration({ start: 0, end: FIVE_MINUTES_IN_MS });
     const otpTTL = formatDuration(otpTTLDuration);
 
-    // if (validOtpEvent.otpType === NotificationType.PHONE_VERIFICATION) {
-    //   throw new Error("Phone number verification is not currently supported.");
-    // }
-
     let message: string = success.OTP_SENT;
-    const user = await UserService.getUserByEmail(validOtpEvent.email);
+    const user = await UserService.getUserByEmail(email);
 
-    if (otpType === NotificationType.EMAIL_VERIFICATION_OTP) {
+    if (eventType === otpEvents.enum.EMAIL_VERIFICATION) {
       if (user.isEmailVerified) throw new Error(ErrorTypeEnum.enum.EMAIL_ALREADY_VERIFIED);
 
       message = success.VERIFICATION_EMAIL_SENT;
 
       await notificationService.sendEmail({
         to: user.email,
-        eventType: otpType,
+        eventType: emailNotificationEvents.enum.EMAIL_VERIFICATION_OTP,
         payload: { otp, recipientName: user.username, otpTTL },
         subject: "Email Verification OTP",
       });
     }
 
-    if (otpType === NotificationType.FORGET_PASSWORD_OTP) {
+    if (eventType === otpEvents.enum.FORGET_PASSWORD) {
       message = success.FORGET_PASSWORD_EMAIL_SENT;
 
       await notificationService.sendEmail({
         to: user.email,
-        eventType: otpType,
+        eventType: emailNotificationEvents.enum.FORGET_PASSWORD_OTP,
         payload: { otp, recipientName: user.username, otpTTL },
         subject: "Forget Password OTP",
       });
@@ -65,19 +61,19 @@ export class OtpService {
     await OtpDAL.upsertOTP({
       userId: user.id,
       otp,
-      otpType,
+      eventType,
       expiryTime: new Date(Date.now() + FIVE_MINUTES_IN_MS),
     });
 
     // Generate a temporary token for the user
-    const actionToken = await new TokenService().createActionToken(user.id, notificationToActionMap[otpType]);
+    const actionToken = await new TokenService().createActionToken(user.id, sendOtpToActionMap[eventType]);
 
     return { message, actionToken };
   }
 
-  public static async verifyOtp(otpEvents: VerifyOtpRequest): Promise<{ message: string; actionToken: string }> {
+  public static async verifyOtp(verifyOtpRequest: VerifyOtpRequest): Promise<{ message: string; actionToken: string }> {
     // Extract the token from the parsed event data
-    const { otp, otpType, actionToken: verificationToken } = verifyOtpRequestSchema.parse(otpEvents);
+    const { otp, eventType, actionToken: verificationToken } = verifyOtpRequestSchema.parse(verifyOtpRequest);
     let actionToken: string = "";
     let message: string = success.OTP_VERIFIED_SUCCESSFULLY;
 
@@ -87,15 +83,15 @@ export class OtpService {
     const tokenService = new TokenService();
     const { userId, id: tokenId } = await tokenService.verifyActionToken(
       verificationToken,
-      notificationToActionMap[otpType] // Dynamically determine the action type based on otpType
+      sendOtpToActionMap[eventType] // Dynamically determine the action type based on eventType
     );
 
     const otpDetails = await OtpService.getOtpDetailsByUserId({
       userId,
-      otpType,
+      eventType,
     });
 
-    const otpData = otpDetails.find((data) => data.otpType === otpType);
+    const otpData = otpDetails.find((data) => data.eventType === eventType);
 
     if (!otpData) throw new Error(ErrorTypeEnum.enum.OTP_NOT_REQUESTED);
 
@@ -105,12 +101,12 @@ export class OtpService {
 
     if (isPast(otpData.expiryTime)) throw new Error(ErrorTypeEnum.enum.OTP_EXPIRED);
 
-    if (otpType === NotificationType.EMAIL_VERIFICATION_OTP) {
+    if (eventType === otpEvents.enum.EMAIL_VERIFICATION) {
       await UserService.updateUser(userId, { isEmailVerified: true });
       message = success.EMAIL_VERIFIED_SUCCESSFULLY;
-      actionToken = await tokenService.createActionToken(userId, UserActionType.AUTO_LOGIN);
-    } else if (otpType === NotificationType.FORGET_PASSWORD_OTP) {
-      actionToken = await tokenService.createActionToken(userId, UserActionType.RESET_PASSWORD);
+      actionToken = await tokenService.createActionToken(userId, userActions.enum.AUTO_LOGIN);
+    } else if (eventType === otpEvents.enum.FORGET_PASSWORD) {
+      actionToken = await tokenService.createActionToken(userId, userActions.enum.RESET_PASSWORD);
       message = success.PASSWORD_RESET_SUCCESSFULLY;
     }